+++ fix_mass/one_img_cat/skiped1.py
@@ -14,7 +14,7 @@ python3 core8/pwb.py fix_mass/one_img_cat/skiped ask
 import sys
 from pathlib import Path
 from api_bots import printe
-from api_bots.ncc_page import ncc_MainPage
+from api_bots.page_ncc import ncc_MainPage
 
 from fix_sets.bots2.text_cat_bot import add_cat_to_set
 from fix_sets.bots2.filter_ids import filter_no_title
