+++ fix_sets/bots2/move_files2.py
@@ -7,7 +7,7 @@ import re
 # import json
 import sys
 from api_bots import printe
-from api_bots.ncc_page import NEW_API
+from api_bots.page_ncc import NEW_API
 from fix_sets.jsons_dirs import get_study_dir#, jsons_dir
 from logs_fix.files import move_text_dir
