+++ fix_sets/ncc_api.py
@@ -5,9 +5,9 @@ from fix_sets.ncc_api import CatDepth
 from fix_sets.ncc_api import post_ncc_params
 
 """
-from api_bots.ncc_page import NEW_API
-from api_bots.ncc_page import CatDepth
-from api_bots.ncc_page import ncc_MainPage
+from api_bots.page_ncc import NEW_API
+from api_bots.page_ncc import CatDepth
+from api_bots.page_ncc import ncc_MainPage
 
 api_new = NEW_API()
