+++ mass/atlas/add_cat.py
@@ -8,7 +8,7 @@ import json
 from pathlib import Path
 from tqdm import tqdm
 from api_bots import printe
-from api_bots.ncc_page import ncc_MainPage
+from api_bots.page_ncc import ncc_MainPage
 
 # Specify the root folder
 main_dir = Path(__file__).parent
