+++ mass/atlas/all_cats.py
@@ -5,7 +5,7 @@ python3 core8/pwb.py atlas/all_cats
 """
 import os
 import json
-from api_bots.ncc_page import ncc_MainPage
+from api_bots.page_ncc import ncc_MainPage
 from images_path import atlas_images_path
 
 root_folder = atlas_images_path
