+++ mass/eyerounds/up.py
@@ -23,7 +23,7 @@ from pathlib import Path
 from nccommons import api
 
 from api_bots import printe
-from api_bots.ncc_page import CatDepth
+from api_bots.page_ncc import CatDepth
 
 from mass.eyerounds.bots.catbot import category_name
 from mass.eyerounds.bots.url_to_title import urls_to_title
