+++ mass/radio/authors_list/save.py
@@ -6,7 +6,7 @@ $HOME/local/bin/python3 core8/pwb.py mass/radio/authors_list/save
 import json
 from pathlib import Path
 
-from api_bots.ncc_page import ncc_MainPage
+from api_bots.page_ncc import ncc_MainPage
 
 main_dir = Path(__file__).parent.parent
 # ---
