+++ mass/radio/authors_list/usa_images.py
@@ -14,8 +14,8 @@ from mass.radio.authors_list.usa import get_usa_auths
 # from mass.radio.lists.cases_to_cats import cases_cats# cases_cats()
 # from mass.radio.authors_list import auths_cats
 
-from api_bots.ncc_page import ncc_MainPage
-from api_bots.ncc_page import CatDepth
+from api_bots.page_ncc import ncc_MainPage
+from api_bots.page_ncc import CatDepth
 from api_bots import printe
 
 # from mass.radio.lists.PD_medical import PD_medical_pages_def
