+++ mass/radio/bots/fix_dup.py
@@ -4,7 +4,7 @@ tfj run fixdup1 --image python3.9 --command "$HOME/local/bin/python3 core8/pwb.p
 
 """
 from api_bots import printe
-from api_bots.ncc_page import ncc_MainPage, CatDepth
+from api_bots.page_ncc import ncc_MainPage, CatDepth
 from mass.radio.jsons_files import jsons
 
 # dumps_jsons(infos=0, urls=0, cases_in_ids=0, cases_dup=0, authors=0, to_work=0, all_ids=0, urls_to_get_info=0)
