+++ mass/radio/lists/cases_to_cats.py
@@ -8,7 +8,7 @@ import json
 import re
 import os
 from datetime import datetime
-from api_bots.ncc_page import CatDepth
+from api_bots.page_ncc import CatDepth
 from api_bots import printe
 from mass.radio.jsons_bot import radio_jsons_dir
