+++ mass/radio/st3o/oao.py
@@ -17,7 +17,7 @@ import sys
 sys.argv.append("del2")
 
 from api_bots import printe
-from api_bots.ncc_page import ncc_MainPage, CatDepth
+from api_bots.page_ncc import ncc_MainPage, CatDepth
 from mass.radio.jsons_files import jsons
 from mass.radio.bots.add_cat import add_cat_to_images
