+++ mass/usaid/all_cats.py
@@ -7,7 +7,7 @@ python3 core8/pwb.py mass/usaid/all_cats ask
 import json
 from pathlib import Path
 from api_bots import printe
-from api_bots.ncc_page import ncc_MainPage
+from api_bots.page_ncc import ncc_MainPage
 
 # Specify the root folder
 main_dir = Path(__file__).parent
