+++ mass/usask/up.py
@@ -20,8 +20,8 @@ from tqdm import tqdm
 from api_bots import printe
 from pathlib import Path
 from nccommons import api
-from api_bots.ncc_page import CatDepth
-from api_bots.ncc_page import ncc_MainPage
+from api_bots.page_ncc import CatDepth
+from api_bots.page_ncc import ncc_MainPage
 
 # Specify the root folder
 main_dir = Path(__file__).parent
