+++ ncc_core/api_bots/page_md.py
@@ -1,6 +1,6 @@
 """
 
-from api_bots.mdwiki_page import load_main_api
+from api_bots.page_md import load_main_api
 
 """
 # ---
