+++ ncc_core/import_to_ncc/bot.py
@@ -5,7 +5,7 @@ python3 core8/pwb.py import_to_ncc/bot
 """
 from api_bots import printe
 from nccommons import api
-from api_bots.ncc_page import ncc_NEW_API
+from api_bots.page_ncc import ncc_NEW_API
 from api_bots.wiki_page import load_main_api
 
 api_new = ncc_NEW_API()
