+++ ncc_core/nc_import/import_bots/import_files.py
@@ -7,7 +7,7 @@ import re
 from . import printe
 from . import upload_file
 from .db import add_to_db, add_to_jsonl
-from .ncc_page import load_main_api
+from .page_ncc import load_main_api
 
 
 def get_file_text(title):
