+++ ncc_core/nccommons/api.py
@@ -19,7 +19,7 @@ import time
 # ---
 from api_bots import printe
 from nccommons import api_upload
-from api_bots.ncc_page import NEW_API
+from api_bots.page_ncc import NEW_API
 
 api_new = NEW_API('www', family='nccommons')
 # api_new.Login_to_wiki()
