+++ ncc_core/nccommons/api_upload.py
@@ -15,7 +15,7 @@ import urllib.request
 from nccommons.ext import get_new_ext
 from nccommons import fix_svg
 from api_bots import printe
-from api_bots.ncc_page import NEW_API
+from api_bots.page_ncc import NEW_API
 
 api_new = NEW_API()
 yes_answer = ["y", "a", "", "Y", "A", "all"]
