+++ sets_dbs/file_infos/p2.py
@@ -9,7 +9,7 @@ import sys
 import json
 from pathlib import Path
 from api_bots import printe
-from api_bots.ncc_page import NEW_API
+from api_bots.page_ncc import NEW_API
 
 Dir = Path(__file__).parent
